now drawing a circle for particle region bounds; SSBOs now following RAII
Previously, my SSBO base class did not generate buffers in the
constructor because the default constructor allowed it to be declared as
a global object.  If it is a global variable, then the constructor is
called prior to main(...) starting up, which means that the buffers
would be generated before the OpenGL context is initialized, which means
the program would blow up.  I said, "Bah!" and changed it anyway for the
sake of RAII, which I am trying to follow more of.  Then just declare
any SSBO objects as pointers and create them with "new" after OpenGL is
initialized.  Problem solved, and now no risk of attempting to use the
SSBO object before it is generated.