Whoops.  Particles move correctly now.  No quad tree yet, but I undid a recent mistake.
When changing Init() in main.cpp to account for the
SsboBase::configureCompute(...) parameter change, I accidentally told
the particle to configure itself for the the "particle update" compute
shader under the buffer name "FaceBuffer".  Whoops.  That was left over
from the previous demo.  Corrected name to "ParticleBuffer".  The
particles move now.