The "quad tree reset" and "quad tree populate" shaders are building now
Aside from the cleanup, I had to alter the AddParticleToNode(...)
function a bit because GLSL does not allow recursion.  So I changed it
to a loop.  It is not running, and the SSBOs for the quad tree nodes and
the quad tree faces are not built yet, and the compute controllers for
"quad tree reset" and "quad tree populate" are not made yet either, but
the shaders are done.  Or at least until they don't work properly and I
have to fix them :).