brief excursion into persistant buffer mapping; will be undoing it
I figured out how to use a persistantly mapped buffer for
ComputeParticleUpdate's "active particles" atomic counter copy buffer,
but I couldn't figure out how to get it to work for the original "active
particles" atomic counter.  Regardless, I learned that persistant
mapping is only useful as long as the buffer is mapped.  I will be
needing to copy other atomic counter to the GL_COPY_WRITE_BUFFER, so I
will not be able to leave one buffer mapped forever anyway.  Back to the
way it was.