Lesson learned: Do NOT attempt to create a mutex in the compute shader.
The purpose of shaders is to launch many threads that run identical code
on different data.  They are expected to run over approximately the same
time frame.  All attempts to create a mutex are attempting to force
these shaders to react to one another.  Without a built-in mechanism,
the result is chaos.

Geometry generation is working now, but I'll have to rethink my
"populate quad tree" shader.