Having some trouble getting a couple integer members of Particle to actually show up as integers.
After much debugging and spending some time making an "oh help" post to
the OpenGL subreddit, I have developed a suspicion that something
happens when data is sent to the GPU.  I remember learning about byte
packing into 16byte chunks, but I am starting to think that such byte
packing is automatically jamming the data into floats despite what it
says in the shader's structure.  The "collision count this frame" in the
vertex shader is exactly the byte pattern of a float version of the
value that compute shader set.

But when I read the particle buffer back from a memory-mapped in the
Update() function, the value shows up fine as an int, so I don't know
why the vertex shader is acting this way.